# chore: upgrade typescript to 4.8, improve compile performance