--- a/package.json
+++ b/package.json
@@ -72,17 +72,17 @@
     "ts-node": "^10.9.1",
     "ts-patch": "^2.0.2",
     "typesafe-dynamodb": "0.1.5",
-    "typescript": "^4.8.2",
+    "typescript": "4.8.2",
     "uuid": "^8.3.2"
   },
   "peerDependencies": {
-    "@aws-cdk/aws-appsync-alpha": "*",
     "aws-cdk-lib": "^2.28.1",
     "aws-sdk": "^2",
     "constructs": "^10.0.0",
     "esbuild": "^0.15.6",
     "typesafe-dynamodb": "^0.1.5",
-    "typescript": "^4.7.2"
+    "typescript": "^4.8.2",
+    "@aws-cdk/aws-appsync-alpha": "*"
   },
   "dependencies": {
     "@functionless/ast-reflection": "^0.2.2",