--- a/src/expression.ts
+++ b/src/expression.ts
@@ -771,7 +771,10 @@ export class TaggedTemplateExpr extends BaseExpr<NodeKind.TaggedTemplateExpr> {
  * // TemplateHead is "".
  * ```
  */
-export class TemplateHead extends BaseNode<NodeKind.TemplateHead> {
+export class TemplateHead extends BaseNode<
+  NodeKind.TemplateHead,
+  TemplateExpr
+> {
   readonly nodeKind = "Node";
 
   constructor(
@@ -795,7 +798,7 @@ export class TemplateHead extends BaseNode<NodeKind.TemplateHead> {
  */
 export class TemplateSpan<
   Literal extends TemplateMiddle | TemplateTail = TemplateMiddle | TemplateTail
-> extends BaseNode<NodeKind.TemplateSpan> {
+> extends BaseNode<NodeKind.TemplateSpan, TemplateExpr> {
   readonly nodeKind = "Node";
 
   constructor(
@@ -815,7 +818,10 @@ export class TemplateSpan<
   }
 }
 
-export class TemplateMiddle extends BaseNode<NodeKind.TemplateMiddle> {
+export class TemplateMiddle extends BaseNode<
+  NodeKind.TemplateMiddle,
+  TemplateSpan
+> {
   readonly nodeKind = "Node";
 
   constructor(
@@ -830,7 +836,10 @@ export class TemplateMiddle extends BaseNode<NodeKind.TemplateMiddle> {
   }
 }
 
-export class TemplateTail extends BaseNode<NodeKind.TemplateTail> {
+export class TemplateTail extends BaseNode<
+  NodeKind.TemplateTail,
+  TemplateSpan
+> {
   readonly nodeKind = "Node";
 
   constructor(