--- a/website/package.json
+++ b/website/package.json
@@ -40,7 +40,7 @@
     "react-hook-form": "^7.31.3",
     "typedoc": "^0.23.13",
     "typedoc-plugin-markdown": "^3.12.1",
-    "typescript": "4.7"
+    "typescript": "^4.8.2"
   },
   "browserslist": {
     "production": [