Fixes type issues which broke in TS 4.8.

Resolved a complex type that was causing an extra second of processing.

### Before
Total time:                  16.91s

```ts
    if (this.parent === undefined) {
      return undefined;
    } else if (is(this.parent)) {
      return this.parent;
    } else {
      return this.parent.findParent(is);
    }
```

### After
Total time:                  16.03s

```ts
    if (this.parent === undefined) {
      return undefined;
    } else if (is(this.parent)) {
      return this.parent as unknown as N;
    } else {
      return this.parent.findParent(is);
    }
```

Evaluating `Parent & {} & N` was taking a second

```
Hot Spots
├─ Check file /home/sussmans/functionless/src/node.ts (1741ms)
│  └─ Compare types 7794 and 1985 (1125ms)
│     └─ Compare types 7794 and 9 (1125ms)
│        ├─ {"id":7794,"kind":"Intersection","count":3,"types":[838,33,1979]}
│        │  ├─ {"id":838,"kind":"TypeParameter","name":"Parent","location":{"path":"/home/sussmans/functionless/src/node.ts","line":76,"char":3}}
│        │  ├─ {"id":33,"kind":"Other","display":"{}"}
│        │  └─ {"id":1979,"kind":"TypeParameter","name":"N","location":{"path":"/home/sussmans/functionless/src/node.ts","line":182,"char":21}}
│        └─ {"id":9,"kind":"Intrinsic","name":"undefined"}
└─ Check file /home/sussmans/functionless/src/asl/synth.ts (1267ms)
```